EH-102 обновил удаление сопоставления

+++ notify-test/public/app/bot/repositories/ExtensionRepository.php
@@ -2,6 +2,8 @@
 
 namespace Pcs\Bot\repositories;
 
+use Pcs\Bot\Logger;
+use Pcs\Bot\Models\Chat;
 use Pcs\Bot\Models\Extension;
 use Pcs\Bot\Models\User;
 
@@ -9,21 +11,29 @@ class ExtensionRepository
 {
     public function getExtensionByExtension($extension)
     {
-        return Extension::all()->where('extension', '=', $extension)->first();
+        return Extension::where('extension', '=', $extension)->first();
     }
 
     public function deleteExtension($extension)
     {
-        $delExtension = Extension::all()->where('extension', '=', $extension)->first();
-        $user = User::all()->where('id', '=', $delExtension->user_id)->first();
+        $delExtension = Extension::where('extension', '=', $extension)->first();
 
-        if ($delExtension->delete()) {
-            $user->delete();
+        if ($delExtension) {
 
-            return true;
-        } else {
-            return null;
+            $chat = Chat::where('user_id', '=', $delExtension->user_id)->first();
+            $user = User::where('id', '=', $delExtension->user_id)->first();
+
+            if ($delExtension->delete()) {
+                if ($chat->delete()) {
+                    $user->delete();
+                    return true;
+                }
+                return false;
+            } else {
+                return null;
+            }
         }
+        return null;
     }
 
     public function getExtensions()
